update Float point
zoom khong con bi sai

+++ src/com/paintjava/List/MyRect.java
@@ -23,10 +23,10 @@ public class MyRect {
 			this.p2=phaitren;
 			this.p3=phaiduoi;
 			this.p4=traiduoi;
-			neo1=new Rectangle(p1.x,p1.y,10,10);
-			neo2=new Rectangle(p2.x,p2.y,10,10);
-			neo3=new Rectangle(p3.x,p3.y,10,10);
-			neo4=new Rectangle(p4.x,p4.y,10,10);
+			neo1=new Rectangle((int)p1.x,(int)p1.y,10,10);
+			neo2=new Rectangle((int)p2.x,(int)p2.y,10,10);
+			neo3=new Rectangle((int)p3.x,(int)p3.y,10,10);
+			neo4=new Rectangle((int)p4.x,(int)p4.y,10,10);
 	}
 	public void Tinhtien(Point t,Point r)
 	{
@@ -34,10 +34,10 @@ public class MyRect {
 		p2.Tinhtien(t,r);
 		p3.Tinhtien(t,r);
 		p4.Tinhtien(t,r);
-		neo1=new Rectangle(p1.x,p1.y,10,10);
-		neo2=new Rectangle(p2.x,p2.y,10,10);
-		neo3=new Rectangle(p3.x,p3.y,10,10);
-		neo4=new Rectangle(p4.x,p4.y,10,10);
+		neo1=new Rectangle((int)p1.x,(int)p1.y,10,10);
+		neo2=new Rectangle((int)p2.x,(int)p2.y,10,10);
+		neo3=new Rectangle((int)p3.x,(int)p3.y,10,10);
+		neo4=new Rectangle((int)p4.x,(int)p4.y,10,10);
 	}
 	public void Zoom(float k)
 	{
